ide_engine/pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "ide-agentic-engine"
version = "2.0.0"
description = "CLI Agentic Engine for IDE AI workflows — optimised for low token spend"
readme = "README.md"
requires-python = ">=3.10"
license = { text = "MIT" }
authors = [{ name = "ide-agentic-engine contributors" }]

dependencies = [
    "anthropic>=0.40.0",
    "instructor[anthropic]>=1.4.0",
    "jsonref>=1.1.0",
    "langgraph>=0.2.0",
    "langchain-core>=0.3.0",
    "pydantic>=2.6.0",
    "redis[hiredis]>=5.0.0",
    "typer>=0.12.0",
    "rich>=13.7.0",
    "opentelemetry-sdk>=1.24.0",
    "opentelemetry-exporter-otlp-proto-grpc>=1.24.0",
    "opentelemetry-api>=1.24.0",
    "sentry-sdk[logging]>=2.0.0",
    "httpx>=0.27.0",
    "python-dotenv>=1.0.0",
    "reportlab>=4.1.0",
    "google-genai>=1.73.1",
    "openai>=2.32.0",
    "modelcontextprotocol>=1.0.1",
    "pyyaml>=6.0.3",
    "fastapi>=0.136.1",
    "uvicorn>=0.46.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.1.0",
    "pytest-asyncio>=0.23.0",
    "pytest-cov>=5.0.0",
    "black>=24.4.0",
    "isort>=5.13.0",
    "flake8>=7.0.0",
    "mypy>=1.9.0",
    "detect-secrets>=1.4.0",
    "pre-commit>=3.7.0",
]

[project.scripts]
ide-agent = "engine.cli:app"

[tool.black]
line-length = 100
target-version = ["py310", "py311", "py312"]

[tool.isort]
profile = "black"
line_length = 100

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
addopts = "-v --tb=short"

[tool.mypy]
python_version = "3.10"
strict = true
ignore_missing_imports = true
exclude = ["tests/"]

[tool.coverage.run]
source = ["engine", "agents", "core", "socroot", "tools"]
omit = ["tests/*", "*/__init__.py"]

[tool.hatch.build.targets.wheel]
packages = ["engine", "agents", "core", "tools", "socroot"]

[dependency-groups]
dev = [
    "pytest-asyncio>=1.3.0",
]
